% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_tables.R
\name{combine_tables}
\alias{combine_tables}
\title{A helper function to combine two projection tables}
\usage{
combine_tables(base_model_table, cop_model_table, assess_year)
}
\arguments{
\item{base_model_table}{Data frame created by \code{\link[=get_model_values]{get_model_values()}}. This should
be the base model.}

\item{cop_model_table}{Data frame created by \code{\link[=get_model_values]{get_model_values()}}. This should be the
catch-only projection model.}

\item{assess_year}{A single numeric value that corresponds to the assessment
year of the model}
}
\description{
A helper function to combine two projection tables
}
\examples{
\dontrun{
base_model <- r4ss::SS_output("C:/model_directory/base_model")
cop_model <- r4ss::SS_output("C:/model_directory/cop_model")

projection_table_base_model <- get_model_values(
  model = base_model
)

projection_table_cop <- get_model_values(
  model = cop_model
)
output <- combine_tables(
  base_model_table = projection_table_base_model,
  cop_model_table = projection_table_cop,
  assess_year = 2023
)
}

}
\author{
Chantel Wetzel, Ian Taylor, Brian Langseth
}
