% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output_table.R
\name{get_output_table}
\alias{get_output_table}
\title{Pull in model results and format into a data table}
\usage{
get_output_table(dir)
}
\arguments{
\item{dir}{Model directory for the model to summarize. This should be a
directory that contains Stock Synthesis model files.}
}
\description{
Create and format a projection table using created by \code{\link[r4ss:SS_output]{r4ss::SS_output()}}
and \code{\link[=get_model_values]{get_model_values()}}.
}
\examples{
\dontrun{
  output_table <- get_output_table(
    dir = "C:/assessments/2027/my_model"
}

}
\author{
Chantel Wetzel
}
