% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_tables.R
\name{combine_tables}
\alias{combine_tables}
\title{A helper function to combine two projection tables}
\usage{
combine_tables(base_model_table, cop_model_table, assess_year)
}
\arguments{
\item{base_model_table}{Data frame created by \code{\link[=get_model_values]{get_model_values()}}. This should
be the base model.}

\item{assess_year}{A single numeric value that corresponds to the assessment
year of the model}

\item{cop_table}{Data frame created by \code{\link[=get_model_values]{get_model_values()}}. This should be the
catch-only projection model.}
}
\description{
A helper function to combine two projection tables
}
\author{
Chantel Wetzel, Ian Taylor, Brian Langseth
}
