% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table.R
\name{format_table}
\alias{format_table}
\title{Format a projection table for a pdf report created using \code{quarto}}
\usage{
format_table(
  table,
  assess_year = NULL,
  sb_decimal_number = 2,
  hl_decimal_number = 0
)
}
\arguments{
\item{table}{Data frame created by \code{\link[=get_model_values]{get_model_values()}} or \code{\link[=combine_tables]{combine_tables()}}
to apply formatting to for reporting.}

\item{assess_year}{A single numeric value that corresponds to the assessment
year of the model. Default is NULL.}

\item{sb_decimal_number}{A single numeric value to determine the number of decimals
to include for spawning biomass/output reporting in the formatted table object.
The default is 2.}

\item{hl_decimal_number}{A single numeric value to determine the number of decimals
to include for OFL, ABC, and ACL reporting in the formatted table object. The
default is 0.}
}
\description{
Create and format a projection table using created by \code{\link[=get_model_values]{get_model_values()}}
or \code{\link[=combine_tables]{combine_tables()}}.
}
\author{
Chantel Wetzel, Ian Taylor, Brian Langseth
}
