% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table.R
\name{format_table}
\alias{format_table}
\title{Format a projection table for a pdf report created using \code{quarto}}
\usage{
format_table(
  table,
  assess_year = NULL,
  sb_decimal_number = 2,
  hl_decimal_number = 0
)
}
\arguments{
\item{table}{Data frame created by \code{\link[=get_model_values]{get_model_values()}} or \code{\link[=combine_tables]{combine_tables()}}
to apply formatting to for reporting.}

\item{assess_year}{A single numeric value that corresponds to the assessment
year of the model. Default is NULL.}

\item{sb_decimal_number}{A single numeric value to determine the number of decimals
to include for spawning biomass/output reporting in the formatted table object.
The default is 2.}

\item{hl_decimal_number}{A single numeric value to determine the number of decimals
to include for OFL, ABC, and ACL reporting in the formatted table object. The
default is 0.}
}
\description{
Create and format a projection table using created by \code{\link[=get_model_values]{get_model_values()}}
or \code{\link[=combine_tables]{combine_tables()}}.
}
\examples{
\dontrun{
base_model <- r4ss::SS_output("C:/model_directory/base_model")
cop_model <- r4ss::SS_output("C:/model_directory/cop_model")

projection_table_base_model <- get_model_values(
  model = base_model
)

projection_table_cop <- get_model_values(
  model = cop_model
)
output <- combine_tables(
  base_model_table = projection_table_base_model,
  cop_model_table = projection_table_cop,
  assess_year = 2023
)

format_table(
  table = output
)
}


}
\author{
Chantel Wetzel, Ian Taylor, Brian Langseth
}
