% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_values.R
\name{get_model_values}
\alias{get_model_values}
\title{Grab and combine model quantities for reporting}
\usage{
get_model_values(replist, buffer = NULL)
}
\arguments{
\item{replist}{Model list object created by \code{\link[r4ss:SS_output]{r4ss::SS_output()}} to summarize
values from.}

\item{buffer}{}
}
\value{
data frame
}
\description{
Grab and combine model quantities for reporting
}
\examples{
\dontrun{
model_output <- r4ss::SS_output("C:/model_directory")
projection_table <- get_model_values(
  replist = model_output
)
}
}
\author{
Chantel Wetzel, Ian Taylor, Brian Langseth
}
